Migrate func_w0611.py to new functional tests
