--- tests/input/func_w0611.py
@@ -1,25 +1,0 @@
-"""check unused import
-"""
-# pylint: disable=no-absolute-import, useless-object-inheritance
-
-from __future__ import print_function
-
-import os
-import sys
-
-class NonRegr(object):
-    """???"""
-    def __init__(self):
-        print('initialized')
-
-    def sys(self):
-        """should not get sys from there..."""
-        print(self, sys)
-
-    def dummy(self, truc):
-        """yo"""
-        return self, truc
-
-    def blop(self):
-        """yo"""
-        print(self, 'blip')
