--- tests/messages/func_w0611.txt
@@ -1,1 +1,0 @@
-W:  7: Unused import os
